Connection to Database
Gotta be able to start the server.js file automatically somehow, prob just on page load.

Also need to update password in server.js file.

In the html-routes.js file (in routes folder), there's a function to run a query, it takes in the query, and returns an array of the results.